requirements.txt
django>=3.2,<5.0
djangorestframework>=3.13
boto3>=1.20